refactor : Result.java
 - Result enum을 패키지로 분리
