+++ src/main/java/domain/result/Result.java
@@ -1,7 +1,7 @@
-package domain;
+package domain.result;
 
 public enum Result {
     승,
     패,
-    무;
+    무
 }
